pyproject.toml
[tool.poetry]
name = "twitter-extractor"
version = "0.1.0"
description = ""
authors = ["Johnny Dang <mail@jydago.com>"]

[tool.poetry.dependencies]
python = "^3.10"
twarc = "^2.10.2"
polars = "^0.13.23"
loguru = "^0.6.0"
python-dotenv = "^0.20.0"
ujson = "^5.2.0"
json-lines = "^0.5.0"
python-dateutil = "^2.8.2"
psycopg2-binary = "^2.9.3"
connectorx = "^0.2.5"

[tool.poetry.dev-dependencies]
pytest = "^7.1.2"
jupyter = "^1.0.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
pythonpath = [
  "."
]
